wip move zit to its own file _zit.py
